requirements.txt
alembic==1.1.0.dev0
apturl==0.5.2
asgiref==3.2.10
Babel==2.6.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.2
django-allauth==0.43.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.18.2
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.4.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
passlib==1.7.2
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.2
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
speaklater==1.4
SQLAlchemy==1.3.12
sqlparse==0.2.4
sshtunnel==0.1.4
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==0.16.1
whitenoise==5.2.0
WTForms==2.2.1
xkit==0.0.0
